+++ packages/odata2ts/src/data-model/DataModelDigestion.ts
@@ -2,7 +2,7 @@ import { MappedConverterChains } from "@odata2ts/converter-runtime";
 
 import { DigestionOptions } from "../FactoryFunctionModel";
 import { DataModel } from "./DataModel";
-import { ComplexType as ComplexModelType, DataTypes, ODataVersion, PropertyModel } from "./DataTypeModel";
+import { ComplexType as ComplexModelType, DataTypes, ModelType, ODataVersion, PropertyModel } from "./DataTypeModel";
 import { ComplexType, EntityType, Property, Schema } from "./edmx/ODataEdmxModelBase";
 import { NamingHelper } from "./NamingHelper";
 import { ServiceConfigHelper } from "./ServiceConfigHelper";
@@ -90,9 +90,12 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
   }
 
   private getBaseModel(model: ComplexType) {
-    const name = this.namingHelper.getModelName(model.$.Name);
-    const qName = this.namingHelper.getQName(model.$.Name);
-    const editableName = this.namingHelper.getEditableModelName(model.$.Name);
+    const entityConfig = this.serviceConfigHelper.findConfigEntityByName(model.$.Name);
+    const entityName = entityConfig?.mappedName || model.$.Name;
+
+    const name = this.namingHelper.getModelName(entityName);
+    const qName = this.namingHelper.getQName(entityName);
+    const editableName = this.namingHelper.getEditableModelName(entityName);
     const odataName = model.$.Name;
     const bType = model.$.BaseType;
     const props = [...(model.Property ?? []), ...this.getNavigationProps(model)];
@@ -132,20 +135,26 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
 
     for (const model of models) {
       const baseModel = this.getBaseModel(model);
+      const entityConfig = this.serviceConfigHelper.findConfigEntityByName(model.$.Name);
+      const entityName = entityConfig?.mappedName || model.$.Name;
 
       // key support: we add keys from this entity,
       // but not keys stemming from base classes (postprocess required)
       const keyNames: Array<string> = [];
-      const entity = model as EntityType;
-      if (entity.Key && entity.Key.length && entity.Key[0].PropertyRef.length) {
-        const propNames = entity.Key[0].PropertyRef.map((key) => key.$.Name);
-        keyNames.push(...propNames);
+      if (entityConfig?.keys?.length) {
+        keyNames.push(...entityConfig.keys);
+      } else {
+        const entity = model as EntityType;
+        if (entity.Key && entity.Key.length && entity.Key[0].PropertyRef.length) {
+          const propNames = entity.Key[0].PropertyRef.map((key) => key.$.Name);
+          keyNames.push(...propNames);
+        }
       }
 
       this.dataModel.addModel(baseModel.name, {
         ...baseModel,
-        idModelName: this.namingHelper.getIdModelName(model.$.Name),
-        qIdFunctionName: this.namingHelper.getQIdFunctionName(model.$.Name),
+        idModelName: this.namingHelper.getIdModelName(entityName),
+        qIdFunctionName: this.namingHelper.getQIdFunctionName(entityName),
         generateId: true,
         keyNames: keyNames, // postprocess required to include key specs from base classes
         keys: [], // postprocess required to include props from base classes
@@ -156,13 +165,15 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
 
   private postProcessModel() {
     // complex types
-    this.dataModel.getComplexTypes().forEach((model) => {
-      const [baseProps] = this.collectBaseClassPropsAndKeys(model);
+    const complexTypes = this.dataModel.getComplexTypes();
+    complexTypes.forEach((model) => {
+      const [baseProps] = this.collectBaseClassPropsAndKeys(model, []);
       model.baseProps = baseProps;
     });
+    const modelTypes = this.dataModel.getModels();
     // entity types
-    this.dataModel.getModels().forEach((model) => {
-      const [baseProps, baseKeys, idName, qIdName] = this.collectBaseClassPropsAndKeys(model);
+    modelTypes.forEach((model) => {
+      const [baseProps, baseKeys, idName, qIdName] = this.collectBaseClassPropsAndKeys(model, []);
       model.baseProps = baseProps;
 
       if (!model.keyNames.length && idName) {
@@ -191,9 +202,55 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
         return prop;
       });
     });
+
+    const sortedModelTypes = this.sortModelsByInheritance<ModelType>(modelTypes);
+    this.dataModel.setModels(sortedModelTypes);
+
+    const sortedComplexTypes = this.sortModelsByInheritance<ComplexModelType>(complexTypes);
+    this.dataModel.setComplexTypes(sortedComplexTypes);
   }
 
-  private collectBaseClassPropsAndKeys(model: ComplexModelType): [Array<PropertyModel>, Array<string>, string, string] {
+  private sortModelsByInheritance<Type extends ComplexModelType>(models: Type[]): { [name: string]: Type } {
+    // recursively visit all models and sort them by inheritance such that base classes
+    // are always before derived classes
+    const sortedModels: { [name: string]: Type } = {};
+    const visitedModels = new Set<Type>();
+    const inProgressModels = new Set<Type>();
+
+    function visit(model: Type) {
+      if (inProgressModels.has(model)) {
+        throw new Error("Cyclic dependencies detected!");
+      }
+
+      if (!visitedModels.has(model)) {
+        inProgressModels.add(model);
+
+        for (const baseClassName of model.baseClasses) {
+          const baseClass = models.find((e) => e.name === baseClassName);
+          if (baseClass) {
+            visit(baseClass);
+          }
+        }
+        visitedModels.add(model);
+        inProgressModels.delete(model);
+        sortedModels[model.name] = model;
+      }
+    }
+
+    for (const model of models) {
+      visit(model);
+    }
+    return sortedModels;
+  }
+
+  private collectBaseClassPropsAndKeys(
+    model: ComplexModelType,
+    visitedModels: string[]
+  ): [Array<PropertyModel>, Array<string>, string, string] {
+    if (visitedModels.includes(model.name)) {
+      throw new Error(`Cyclic inheritance detected for model ${model.name}!`);
+    }
+    visitedModels.push(model.name);
     return model.baseClasses.reduce(
       ([props, keys, idName, qIdName], bc) => {
         const baseModel = this.dataModel.getModel(bc) || this.dataModel.getComplexType(bc);
@@ -202,7 +259,10 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
 
         // recursive
         if (baseModel.baseClasses.length) {
-          const [parentProps, parentKeys, parentIdName, parentQIdName] = this.collectBaseClassPropsAndKeys(baseModel);
+          const [parentProps, parentKeys, parentIdName, parentQIdName] = this.collectBaseClassPropsAndKeys(
+            baseModel,
+            visitedModels
+          );
           props.unshift(...parentProps);
           keys.unshift(...parentKeys);
           if (parentIdName) {
@@ -212,7 +272,7 @@ export abstract class Digester<S extends Schema<ET, CT>, ET extends EntityType, 
         }
 
         props.push(...baseModel.props);
-        if (baseModel.keyNames) {
+        if (baseModel.keyNames?.length) {
           keys.push(...baseModel.keyNames.filter((kn) => !keys.includes(kn)));
           idNameResult = baseModel.idModelName;
           qIdNameResult = baseModel.qIdFunctionName;