+++ packages/odata2ts/test/data-model/DataModelDigestionTests.ts
@@ -64,6 +64,7 @@ export function createDataModelTests(
 
   test("using Id of base class", async () => {
     odataBuilder
+      .addEntityType("Child", "Parent", () => {})
       .addEntityType("GrandParent", undefined, (builder) => {
         builder.addKeyProp("ID", "Edm.String");
       })
@@ -71,11 +72,11 @@ export function createDataModelTests(
 
     const result = await doDigest();
 
-    expect(result.getModels().length).toBe(2);
-    expect(result.getModels()[1].name).toBe("Parent");
-    expect(result.getModels()[1].idModelName).toBe("GrandParentId");
-    expect(result.getModels()[1].qIdFunctionName).toBe("QGrandParentId");
-    expect(result.getModels()[1].generateId).toBe(false);
+    expect(result.getModels().length).toBe(3);
+    expect(result.getModels()[2].name).toBe("Child");
+    expect(result.getModels()[2].idModelName).toBe("GrandParentId");
+    expect(result.getModels()[2].qIdFunctionName).toBe("QGrandParentId");
+    expect(result.getModels()[2].generateId).toBe(false);
   });
 
   test("complex Id with base class", async () => {
@@ -96,6 +97,37 @@ export function createDataModelTests(
     expect(result.getModels()[1].idModelName).toBe("ParentId");
     expect(result.getModels()[1].qIdFunctionName).toBe("QParentId");
     expect(result.getModels()[1].generateId).toBe(true);
+  });
+
+  test(`base classes with cyclical dependencies`, async () => {
+    expect.assertions(1);
+    odataBuilder
+      .addEntityType("Child", "Parent", (builder) => builder)
+      .addEntityType("Parent", "Child", (builder) => builder);
+
+    await expect(doDigest()).rejects.toThrowError("Cyclic inheritance detected for model Child!");
+  });
+
+  test(`reordering of classes by inheritance`, async () => {
+    odataBuilder
+      .addEntityType("GrandChild", "Child", (builder) => builder)
+      .addEntityType("Child", "Parent", (builder) => builder)
+      .addEntityType("StandAlone", undefined, (builder) => {
+        builder.addKeyProp("ID", "Edm.String");
+      })
+      .addEntityType("GrandParent", undefined, (builder) => {
+        builder.addKeyProp("ID", "Edm.String");
+      })
+      .addEntityType("Parent", "GrandParent", (builder) => builder);
+
+    const result = await doDigest();
+
+    expect(result.getModels().length).toBe(5);
+    expect(result.getModels()[0].name).toBe("GrandParent");
+    expect(result.getModels()[1].name).toBe("Parent");
+    expect(result.getModels()[2].name).toBe("Child");
+    expect(result.getModels()[3].name).toBe("GrandChild");
+    expect(result.getModels()[4].name).toBe("StandAlone");
   });
 
   test.skip("converter test", async () => {
@@ -179,5 +211,66 @@ export function createDataModelTests(
     expect(toTest.idModelName).toBe("TEST_KEY_MODEL");
     expect(toTest.qIdFunctionName).toBe("YYY_TEST_KEY_FUNC");
     expect(toTest.editableName).toBe("TEST_EDIT_DUMMY");
+  });
+
+  test("property configuration", async () => {
+    digestionOptions.allowRenaming = true;
+    digestionOptions.propertiesByName = [
+      { name: "ID", mappedName: "newId", managed: true },
+      { name: /ageOfEmpire/, mappedName: "age" },
+    ];
+
+    odataBuilder
+      .addEntityType("Test", undefined, (builder) => {
+        builder.addKeyProp("ID", "Edm.String");
+      })
+      .addComplexType("ComplexTest", undefined, (builder) => {
+        builder.addProp("ageOfEmpire", "Edm.Int32");
+      });
+
+    const result = await doDigest();
+
+    let toTest = result.getModels()[0].props[0];
+    expect(toTest.odataName).toBe("ID");
+    expect(toTest.name).toBe("newId");
+    expect(toTest.managed).toBe(true);
+
+    toTest = result.getComplexTypes()[0].props[0];
+    expect(toTest.odataName).toBe("ageOfEmpire");
+    expect(toTest.name).toBe("age");
+    expect(toTest.managed).toBeUndefined();
+  });
+
+  test("entity configuration", async () => {
+    digestionOptions.allowRenaming = true;
+    digestionOptions.entitiesByName = [
+      { name: "Test", mappedName: "newTest", keys: ["ID", "Version"] },
+      { name: /Complex.*/, mappedName: "cmplx" },
+    ];
+
+    odataBuilder
+      .addEntityType("Test", undefined, (builder) => {
+        builder.addKeyProp("ID", "Edm.String");
+        builder.addProp("Version", "Edm.String");
+      })
+      .addComplexType("ComplexTest", undefined, (builder) => {
+        builder.addProp("ageOfEmpire", "Edm.Int32");
+      });
+
+    const result = await doDigest();
+
+    let toTestCmplx = result.getComplexTypes()[0];
+    expect(toTestCmplx.odataName).toBe("ComplexTest");
+    expect(toTestCmplx.name).toBe("Cmplx");
+    expect(toTestCmplx.editableName).toBe("EditableCmplx");
+
+    let toTest = result.getModels()[0];
+    expect(toTest.odataName).toBe("Test");
+    expect(toTest.name).toBe("NewTest");
+    expect(toTest.idModelName).toBe("NewTestId");
+    expect(toTest.qIdFunctionName).toBe("QNewTestId");
+    expect(toTest.editableName).toBe("EditableNewTest");
+    expect(toTest.keyNames).toStrictEqual(["ID", "Version"]);
+    expect(toTest.keys.length).toBe(2);
   });
 }